Extract TFD tests to a traits
